requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
sentence-transformers
transformers
torch
chromadb
guardrails-ai
rouge-score
scikit-learn
gunicorn
pydantic
python-multipart
rich
typer
openai
jsonschema
mlflow
python-dotenv
sqlalchemy
passlib==1.7.4
bcrypt==4.0.1
python-multipart
prometheus-client
python-jose
mcp
fastmcp
celery
redis
httpx
langchain-community
langchain_text_splitters
langchain
unstructured[local-inference]




